fixed bug that caused inPlace to have reverse effect in loadRLObject, updated unit tests
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/recordlinkage@249 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ branches/devel/R/serialization.r
@@ -16,6 +16,7 @@ setMethod(
     result@dbFile <- tempfile()
     sqliteCopyDatabase(from = object@con, to = result@dbFile)
     result@con <- dbConnect(result@drv, dbname = result@dbFile)
+    init_sqlite_extensions(result@con)
     result
   }
 )
@@ -33,37 +34,43 @@ setMethod(
 # additional table to hold the R object
 setGeneric(
   name = "saveRLObject",
-  def = function(object, file, ...) standardGeneric("saveRLObject")
+  def = function(object, file = getDbFile(object), ...) standardGeneric("saveRLObject")
 )
 
 setMethod(
   f = "saveRLObject",
   signature = "RLBigData",
-  definition = function(object, file, ...)
+  definition = function(object, file = getDbFile(object), ...)
   {
+    # only if file differs from current working copy:
     # first, save existing database and open for the following operations
-    sqliteCopyDatabase(from = object@con, to = file)
-    target <- dbConnect(object@drv, file)
+    if (!missing(file))
+    {
+      sqliteCopyDatabase(from = object@con, to = file)
+      target <- dbConnect(object@drv, file)
+    } else target <- object@con
     
     # serialize R object and store in a BLOB
     serializedObj <- serialize(object, connection = NULL)
+    dbGetQuery(target, "drop table if exists serialization")
     dbGetQuery(target, "create table serialization (data blob)")
     dbGetPreparedQuery(target, "insert into serialization values (:val)",
       data.frame(val=I(list(serializedObj))))
-    dbDisconnect(target)
+    # if the target db is newly created, it can be closed
+    if (!missing(file)) dbDisconnect(target)
   }
 )
 
 setMethod(
   f = "saveRLObject",
   signature = "RLResult",
-  definition = function(object, file, ...)
+  definition = function(object, file = getDbFile(object), ...)
   {
   }
 )
 
 # Load object from disk.
-loadRLObject <- function(file)
+loadRLObject <- function(file, inPlace = FALSE)
 {
   # open connection and verify that a table serializatin exists
   # (this is where the object is saved)
@@ -78,12 +85,21 @@ loadRLObject <- function(file)
   if (!inherits(object, c("RLBigData", "RLResult")))
     stop(sprintf("Found an object of unsupported class: %s", class(object)))
 
+  # for inPlace = FALSE:
   # copy database to temporary file and delete serialization in db
-  newfile <- tempfile()
-  sqliteCopyDatabase(from = con, to = newfile)
-  dbDisconnect(con)
-  con <- dbConnect(drv, dbname=newfile)
-  dbGetQuery(con, "drop table serialization")
+  if(!inPlace)
+  {
+    newfile <- tempfile()
+    sqliteCopyDatabase(from = con, to = newfile)
+    dbDisconnect(con)
+    con <- dbConnect(drv, dbname=newfile)
+    dbGetQuery(con, "drop table serialization")
+  } else
+  {
+    # otherwise keep the file and do not delete object so that file can
+    # be reloaded if closed without calling saveRLObject again
+    newfile <- file
+  }
 
   # attach connection info to object
   if(inherits(object, "RLBigData"))
@@ -91,11 +107,30 @@ loadRLObject <- function(file)
     object@dbFile <- newfile
     object@drv <- drv
     object@con <- con
+    init_sqlite_extensions(object@con)
   } else if (inherits(object, "RLResult"))
   {
     object@data@dbFile <- newfile
     object@data@drv <- drv
     object@data@con <- con
+    init_sqlite_extensions(object@data@con)
   }
   object
 }
+
+setGeneric(
+  name = "getDbFile",
+  def = function(object) standardGeneric("getDbFile")
+)
+
+setMethod(
+  f = "getDbFile",
+  signature = "RLBigData",
+  definition = function(object) object@dbFile
+)
+
+setMethod(
+  f = "getDbFile",
+  signature = "RLBigData",
+  definition = function(object) object@data@dbFile
+)
